Create What Must I Do to be Saved?
After carefully looking at Statements of Faith on church websites and asking around, I've noticed that the Church worldwide is not together on this most fundamental and most important doctrine.  Generalizations are made, things are left out or added, and there's a tendency to confuse salvation with personal righteousness or sanctification.  When the church does know it, they often have some wyrde sickness that causes them to withhold it from the World in hopes that their friendship will cause the dying World to ask (but God calls Worldlings to salvation and He is the one who sends the preacher to tell them; also, "friendship with the World is enmity with God").

These things ought not to be, because salvation doctrine is explicit and clear in the Bible as is the fact that salvation is not dependent on our will to keep from sin.

After we are saved, we have a new nature that urges us to not sin and we have power to not sin in a way that we did not have before we were born again, given a new nature and saved.  It's Christ's righteousness that secures our salvation, not ours.